+++ PremiumCalculator/Model/Rating.cs
@@ -9,6 +9,6 @@ namespace PremiumCalculator.Model
     {
         public int ID { get; set; }
         public string Name { get; set; }
-        public double FactorValue { get; set; }
+        public decimal FactorValue { get; set; }
     }
 }
